pyproject.toml
[tool.poetry]
name = "estesis"
version = "0.1.0"
description = ""
authors = ["Pupupupupipu <vika566384@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
redis = "^4.3.4"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.4.0"
flower = "^2.0.1"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
